Code refactoring for improved stability and reuse. Remove context menu.
