+++ source/popup.js
@@ -1,46 +1,43 @@
+import isGitHub from './isGitHub.js';
+import getGitHubLinks from './getGitHubLinks.js';
+
 const linkTarget = document.getElementById('link-target');
 const linkList = document.getElementById('link-list');
 
-chrome.tabs.query({ active: true, currentWindow: true }, function (tabs) {
-    if (tabs && tabs.length == 1) {
-        while (linkList.firstChild) linkList.removeChild(linkList.firstChild);
+chrome.tabs.query({ active: true, currentWindow: true }, ([tab]) => {
+    if (!isGitHub(tab)) return;
 
-        const { id, title, url } = tabs[0];
-        const { origin, pathname, hash } = new URL(url);
-        const [, org, repo, pullOrIssues, number] = pathname.split('/');
+    const {id, url, title} = tab;
 
-        const linkTargetAnchor = document.createElement("A");
-        linkTarget.innerText = url;
+    const linkTargetAnchor = document.createElement("A");
+    linkTarget.innerText = url;
 
-        chrome.tabs.sendMessage(id, { type: 'getLinks' }, links => {
-            if (!chrome.runtime.lastError && links && links.length > 0) {
-                links.filter(l => !l.disabled).forEach(({ text, separator }) => {
-                    if (separator && linkList.lastChild) {
-                        linkList.lastChild.className = 'last-in-group';
-                    }
+    while (linkList.firstChild) linkList.removeChild(linkList.firstChild);
 
-                    if (text) {
-                        const anchor = document.createElement("A");
-                        anchor.innerText = text;
-                        anchor.href = url;
-                        anchor.title = `Click to copy this link to the clipboard.\n\nText:\n${text}\n\nURL:\n${url}`;
+    getGitHubLinks(url, title).filter(l => !l.disabled).forEach(({ text, separator }) => {
+        if (separator && linkList.lastChild) {
+            linkList.lastChild.className = 'copy-github-link-separator';
+        }
 
-                        anchor.onclick = event => {
-                            chrome.tabs.sendMessage(id, { type: "copyLink", url, text });
+        if (text) {
+            const anchor = document.createElement("A");
+            anchor.innerText = text;
+            anchor.href = url;
+            anchor.title = `Click to copy this link to the clipboard.\n\nText:\n${text}\n\nURL:\n${url}`;
 
-                            anchor.className = 'clicked';
-                            window.setTimeout(() => anchor.className = null, 250);
-                            window.setTimeout(() => window.close(), 300);
+            anchor.onclick = event => {
+                chrome.tabs.sendMessage(id, { type: "copyLink", url, text });
 
-                            event.preventDefault();
-                        };
+                anchor.className = 'copy-github-link-clicked';
+                window.setTimeout(() => anchor.className = null, 250);
+                window.setTimeout(() => window.close(), 300);
 
-                        const item = document.createElement("LI");
-                        item.appendChild(anchor);
-                        linkList.appendChild(item);
-                    }
-                });
-            }
-        });
-    }
+                event.preventDefault();
+            };
+
+            const item = document.createElement("LI");
+            item.appendChild(anchor);
+            linkList.appendChild(item);
+        }
+    });
 });
